+++ src/parser/parse.js
@@ -2,14 +2,13 @@
 // @flow
 import path from 'path';
 import {parse} from '@babel/parser';
-import gql from 'graphql-tag';
 import type {
     BabelNodeImportDeclaration,
     BabelNodeVariableDeclarator,
     BabelNodeTaggedTemplateExpression,
+    BabelNodeFile,
 } from '@babel/types';
 import traverse from '@babel/traverse';
-import type {DocumentNode} from 'graphql/language/ast';
 
 /**
  * This file is responsible for finding all gql`-annotated
@@ -38,112 +37,30 @@ import type {DocumentNode} from 'graphql/language/ast';
  *   in `gqlOp<Type>()` or not (to inform an auto-wrapper of the future)
  */
 
-const resolveImport = (
-    expr: Import,
-    files: Files,
-    errors: FileResult['errors'],
-    seen: {[key: string]: true},
-): ?Document => {
-    if (seen[expr.path]) {
-        errors.push({
-            loc: expr.loc,
-            message: `Circular import ${Object.keys(seen).join(' -> ')} -> ${
-                expr.path
-            }`,
-        });
-        return null;
-    }
-    seen[expr.path] = true;
-    const res = files[expr.path];
-    if (!res) {
-        errors.push({loc: expr.loc, message: `No file ${expr.path}`});
-        return null;
-    }
-    if (!res.exports[expr.name]) {
-        errors.push({
-            loc: expr.loc,
-            message: `${expr.path} has no valid gql export ${expr.name}`,
-        });
-        return null;
-    }
-    const value = res.exports[expr.name];
-    if (value.type === 'import') {
-        return resolveImport(value, files, errors, seen);
-    }
-    return value;
-};
-
-const resolveGqlTemplate = (
-    template: Template,
-    files: Files,
-    errors: FileResult['errors'],
-    resolved: Resolved,
-    seen: {[key: string]: true},
-): ?DocumentNode => {
-    const key = template.loc.path + ':' + template.loc.start;
-    if (seen[key]) {
-        errors.push({
-            loc: template.loc,
-            message: `Recursive template dependency! ${Object.keys(seen).join(
-                ' -> ',
-            )} -> ${key}`,
-        });
-        return null;
-    }
-    seen[key] = true;
-    if (resolved[key]) {
-        return resolved[key].document;
-    }
-    const processed: {[key: string]: DocumentNode} = {};
-    const expressions = template.expressions.map((expr) => {
-        if (expr.type === 'import') {
-            const document = resolveImport(expr, files, errors, {});
-            return document
-                ? resolveGqlTemplate(
-                      document.source,
-                      files,
-                      errors,
-                      resolved,
-                      seen,
-                  )
-                : null;
-        }
-        return resolveGqlTemplate(expr.source, files, errors, resolved, seen);
-    });
-    if (expressions.includes(null)) {
-        return null;
-    }
-    resolved[key] = {
-        document: gql(template.literals, ...expressions),
-        raw: template,
-    };
-    return resolved[key].document;
-};
-
-type Template = {|
+export type Template = {|
     literals: Array<string>,
     expressions: Array<Document | Import>,
     loc: Loc,
 |};
-type Loc = {start: number, end: number, path: string};
+export type Loc = {start: number, end: number, path: string};
 
-type Document = {|
+export type Document = {|
     type: 'document',
     source: Template,
 |};
-type Import = {|
+export type Import = {|
     type: 'import',
     name: string,
     path: string,
     loc: Loc,
 |};
 
-type Operation = {|
+export type Operation = {|
     source: Template,
     needsWrapping: boolean,
 |};
 
-type FileResult = {|
+export type FileResult = {|
     path: string,
     operations: Array<Operation>,
     exports: {[key: string]: Document | Import},
@@ -151,7 +68,7 @@ type FileResult = {|
     errors: Array<{loc: Loc, message: string}>,
 |};
 
-type Files = {[path: string]: FileResult};
+export type Files = {[path: string]: FileResult};
 
 const externalReferences = (file: FileResult): Array<string> => {
     const paths = {};
@@ -200,7 +117,7 @@ export const processFile = (filePath: string, contents: string): FileResult => {
         locals: {},
         errors: [],
     };
-    const ast = parse(contents, {
+    const ast: BabelNodeFile = parse(contents, {
         sourceType: 'module',
         allowImportExportEverywhere: true,
         plugins: [['flow', {enums: true}], 'jsx'],
@@ -212,7 +129,7 @@ export const processFile = (filePath: string, contents: string): FileResult => {
         if (toplevel.type === 'ImportDeclaration') {
             const newLocals = getLocals(dir, toplevel, filePath);
             if (newLocals) {
-                for (let k in newLocals) {
+                Object.keys(newLocals).forEach((k) => {
                     const local = newLocals[k];
                     if (local.path.startsWith('/')) {
                         result.locals[k] = local;
@@ -223,21 +140,31 @@ export const processFile = (filePath: string, contents: string): FileResult => {
                     ) {
                         gqlTagNames.push(k);
                     }
-                }
+                });
             }
         }
         if (toplevel.type === 'ExportNamedDeclaration') {
             if (toplevel.source) {
-                const importPath = toplevel.source.value.startsWith('.')
-                    ? path.resolve(path.join(dir, toplevel.source.value))
-                    : toplevel.source.value;
+                const source = toplevel.source;
+                const importPath = source.value.startsWith('.')
+                    ? path.resolve(path.join(dir, source.value))
+                    : source.value;
                 toplevel.specifiers.forEach((spec) => {
-                    result.exports[spec.exported.name] = {
-                        type: 'import',
-                        name: spec.local.name,
-                        path: importPath,
-                        loc: {start: spec.start, end: spec.end, path: filePath},
-                    };
+                    if (
+                        spec.type === 'ExportSpecifier' &&
+                        spec.exported.type === 'Identifier'
+                    ) {
+                        result.exports[spec.exported.name] = {
+                            type: 'import',
+                            name: spec.local.name,
+                            path: importPath,
+                            loc: {
+                                start: spec.start,
+                                end: spec.end,
+                                path: filePath,
+                            },
+                        };
+                    }
                 });
             } else {
                 toplevel.specifiers.forEach((spec) => {
@@ -269,7 +196,6 @@ export const processFile = (filePath: string, contents: string): FileResult => {
                 ) {
                     if (gqlTagNames.includes(init.tag.name)) {
                         const tpl = processTemplate(init, result);
-                        console.log('PROC', init.start);
                         if (tpl) {
                             const document = (result.locals[id] = {
                                 type: 'document',
@@ -427,44 +353,6 @@ const getLocals = (
         }
     });
     return locals;
-};
-type Resolved = {
-    [key: string]: {
-        document: DocumentNode,
-        raw: Template,
-    },
-};
-
-export const resolveDocuments = (
-    files: Files,
-): {resolved: Resolved, errors: FileResult['errors']} => {
-    const resolved: Resolved = {};
-    const errors: FileResult['errors'] = [];
-    Object.keys(files).forEach((path) => {
-        const file = files[path];
-        file.operations.forEach((op) => {
-            const result = resolveGqlTemplate(
-                op.source,
-                files,
-                errors,
-                resolved,
-                {},
-            );
-        });
-        Object.keys(file.locals).forEach((k) => {
-            const local = file.locals[k];
-            if (local.type === 'document') {
-                const result = resolveGqlTemplate(
-                    local.source,
-                    files,
-                    errors,
-                    resolved,
-                    {},
-                );
-            }
-        });
-    });
-    return {resolved, errors};
 };
 
 export const processFiles = (
